utftech/requirements.txt
Django==4.2.14
djangorestframework==3.15.2
django-model-utils==4.5.1
python-dotenv==1.0.1